requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
requests>=2.31.0
cachetools>=5.0.0
pydantic>=1.10.0
openai>=1.33.0
cryptography>=40.0.0
